% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_coloc.R
\name{colocMolecularQTLs}
\alias{colocMolecularQTLs}
\title{Perform colocalisation between an single QTL and GWAS summary stats from the same region.}
\usage{
colocMolecularQTLs(qtl_df, qtl_summary_path, gwas_summary_path,
  variant_info, N_qtl = 84, cis_dist = 1e+05, gwas_type)
}
\arguments{
\item{qtl_df}{A data frame with a single row and two columns (phenotype_id, snp_id) corresponding the a single QTL.}

\item{qtl_summary_path}{Path to the tabix indexed QTL summary statistics file generated by QTLtools}

\item{gwas_summary_path}{Path to the tabix indexed GWAS summary statistics file.}

\item{variant_info}{Variant information for the QTL data. Imported using importVariantInformation() function.}

\item{N_qtl}{Sample size for QTL mapping. Used by coloc to estimate the the standard errors from p-values and effect sizes.}

\item{cis_dist}{With of the genomic region around the lead QTL variant that is used for colocalisation; width = 2*cis_dist.}

\item{gwas_type}{type of GWAS summary stats used for coloc.}
}
\value{
List of colocalisation results or NULL values if there was an error.
}
\description{
Perform colocalisation between an single QTL and GWAS summary stats from the same region.
}
