% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{tabixFetchGWASCatalogSummary}
\alias{tabixFetchGWASCatalogSummary}
\title{Import a specific region from a tabix-indexed GWAS summary stats file}
\usage{
tabixFetchGWASCatalogSummary(granges, summary_path)
}
\description{
Import a specific region from a tabix-indexed GWAS summary stats file
}
